+++ createpythonpackage/__main__.py
@@ -1,0 +1,4 @@
+from .main import create_package
+
+if __name__ == "__main__":
+    create_package()