+++ createpythonpackage/main.py
@@ -2,148 +2,83 @@
 
 import argparse
 import logging
-import os
 from pathlib import Path
-import subprocess
-import sys
-from createpythonpackage import templates
+from .Package import Package, PackageConfig, PackageEnv, PackageLicense
+from .util import blue, printblue, TEST_PYPI_URL, CppError, mkdir, run, grey
+from typing import Optional, List
 
-TEST_PYPI_URL = "https://test.pypi.org/simple/"
-ISATTY = sys.stdout.isatty()
-WINDOWS = os.name == "nt"
-
-__version__ = "0.1.0.0"
-
-
-def blue(text):
-    if ISATTY and not WINDOWS:
-        return "\033[1;34m" + text + "\033[0m"
-    return text
-
-
-def printblue(text):
-    print(blue(text))
+__version__ = "0.2.0.0"
 
 
 def print_version():
     print(__version__)
 
 
-class CppError(Exception):
-    pass
-
-
-def _run(cmd, check=True, **kwargs):
-    cmd_str = " ".join(str(c) for c in cmd)
-    logging.info(f"running {cmd_str}")
-    returncode = subprocess.run(cmd, **kwargs).returncode
-    if check and returncode:
-        raise CppError(f"{cmd_str!r} failed with returncode {str(returncode)}")
-    return returncode
-
-
-def _create_venv(path: Path, name: str, python: str = sys.executable):
-    venv = path / "venv"
-    if venv.exists() and len(list(venv.iterdir())):
-        raise CppError(
-            f"Directory {str(venv)!r} already exists. Remove then try again."
-        )
-    print(f"Creating a virtual environment at {blue(str(venv))}")
-    if _run([python, "-m", "venv", venv, "--prompt", name]):
-        raise CppError("Could not create virtual environment")
-    print(f"Upgrading {blue('pip')} in the virtual environment.")
-    if _run([path / "venv/bin/pip", "install", "--upgrade", "--quiet", "pip"]):
-        logging.warning("Could not upgrade pip to latest verions")
-    print()
-    return venv
-
-
-class Package:
-    def __init__(self, path, name, author, version, email):
-        self.path = path
-        self.name = name
-        self.author = author
-        self.version = version
-        self.email = email
-        self.add_files()
-        _create_venv(self.path, self.name)
-        self.git_init()
-
-    def git_init(self):
-        _run(["git", "init", self.path], stdout=subprocess.DEVNULL)
-        print("Initialized a git repository.")
-        print()
-
-    def add_files(self):
-        (self.path / self.name).mkdir(exist_ok=True)
-        (self.path / self.name / "__init__.py").touch()
-
-        with open(self.path / "README.md", "w") as f:
-            f.write(templates.readme.format(name=self.name, author=self.author))
-
-        with open(self.path / "setup.py", "w") as f:
-            f.write(
-                templates.setup.format(
-                    name=self.name, author=self.author, email=self.email
-                )
-            )
-
-        with open(self.path / self.name / "main.py", "w") as f:
-            f.write(templates.main.format(version=self.version))
-
-        with open(self.path / ".gitignore", "w") as f:
-            f.write(templates.gitignore)
-
-        with open(self.path / "LICENSE", "w") as f:
-            f.write(templates.licence)
-
-        with open(self.path / "makefile", "w") as f:
-            f.write(templates.makefile)
-
-    def print_success(self):
-        print(f"Success! Created {self.name} at {str(self.path)}")
-        print("Inside that directory, you can run several commands")
-        print()
-        printblue("  source ven/bin/activate")
-        print("     Activates this package's isolated Python environment")
-        print()
-        printblue("  pip install PACKAGE")
-        print("    Installs a package to current environment")
-        print()
-        printblue("  pip install -e .")
-        print("    Installs this package in editable mode to the current environment")
-        print()
-        print(
-            f"We suggest that you being by typing:\n\n"
-            f"  {blue('cd')} {self.name}\n"
-            f"  {blue('source ven/bin/activate')}\n\n"
-            "To deactivate the virtual environment, type `deactivate`.\n"
-        )
-        print(
-            "Questions? Create an issue at https://github.com/cs01/create-python-package"
-        )
-        print()
-        print("Happy hacking!")
-
-
-def mkdir(path):
-    if path.is_dir():
-        return
-    logging.info(f"creating directory {path}")
-    path.mkdir(parents=True, exist_ok=True)
+def question(s, default, options: Optional[List[str]] = None):
+    while True:
+        question = grey("question")
+        if options:
+            options_str = ", ".join(options)
+            response = input(f"{question} {s} ({default}) (options: {options_str}): ")
+        else:
+            response = input(f"{question} {s} ({default}): ")
+        if not response:
+            return default
+        return response
 
 
 def _create_package(args):
-    author = args.author
-    version = args.packageversion
-    name = Path(args.name).name
     path = Path(args.name).resolve()
-    if Path(args.name).exists() and len(list(path.iterdir())):
+    if path.exists() and len(list(path.iterdir())) and not args.force:
         raise CppError(f"{str(path)} already exists and is not empty")
+
+    envs = [o.name for o in PackageEnv]
+    licenses = [o.name for o in PackageLicense]
+
+    default_version = "0.0.0.1"
+    default_description = ""
+    default_entrypoint = "main.py"
+    default_repo_url = ""  # TODO get repo origin if in a git repo
+    default_author = "Your Name"  # TODO get repo author if in a git repo
+    default_email = "email@doman.com"  # TODO get repo email if in a git repo
+    default_env = envs[0]
+    default_license = licenses[0]
+
+    if args.yes:
+        package_config = PackageConfig(
+            **{
+                "path": path,
+                "version": default_version,
+                "description": default_description,
+                "entrypoint": default_entrypoint,
+                "repo_url": default_repo_url,
+                "author": default_author,
+                "email": default_email,
+                "env": default_env,
+                "userlicense": default_license,
+                "force": args.force,
+            }
+        )
+    else:
+        package_config = PackageConfig(
+            **{
+                "path": path,
+                "version": question("version", default_version),
+                "description": question("description", default_description),
+                "entrypoint": question("entry point", default_entrypoint),
+                "repo_url": question("repository url", default_repo_url),
+                "author": question("author", default_author),
+                "email": question("email", default_email),
+                "env": question("environment management", default_env, options=envs),
+                "userlicense": question("license", default_license, options=licenses),
+                "force": args.force,
+            }
+        )
+
     print(f"Creating a new Python package in {blue(str(path))}")
     print()
     mkdir(path)
-    return Package(path, name, author, version, args.email)
+    return Package(package_config)
 
 
 def _setup_logging(verbose):
@@ -169,13 +104,13 @@ def _build_package(path):
     cmd = ["pip", "install", "--upgrade", "--quiet", "setuptools", "wheel", "twine"]
     print(f"  {' '.join(list(str(c) for c in cmd))}")
     print()
-    _run(cmd)
+    run(cmd)
 
     cmd = ["python", path / "setup.py", "--quiet", "sdist", "bdist_wheel"]
     printblue("Building package")
     print(f"  {' '.join(list(str(c) for c in cmd))}")
     print()
-    _run(cmd)
+    run(cmd)
     printblue("Package has been built! See output in 'dist' directory.")
 
 
@@ -187,43 +122,41 @@ def _publish(path, test):
     printblue("Uploading package")
     print(f"  {' '.join(list(str(c) for c in cmd))}")
     print()
-    _run(cmd)
+    run(cmd)
 
 
 def create_package():
     parser = argparse.ArgumentParser(
-        description=(
-            "Creates boilerplate for Python projects. Includes setup.py, "
-            "git repository, .gitinore, license, README, makefile."
-        ),
+        description="Create the file and folder structure for a Python package",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
-    parser.add_argument("name", nargs="?", help="Name of package to create")
-    parser.add_argument("--author", help="Name of author", default="Your Name")
+    default_name = Path.cwd().name
     parser.add_argument(
-        "--email", help="Email of author", default="youremail@domain.com"
+        "name", nargs="?", default=default_name, help="Name of package to create"
     )
+    parser.add_argument("--yes", action="store_true", help="Use default options")
     parser.add_argument(
-        "--packageversion", help="Initial version of your package", default="0.0.0.1"
+        "--force",
+        help="create package in existing directory (may delete or overwrite existin files!)",
+        action="store_true",
     )
     parser.add_argument("--version", action="store_true")
     parser.add_argument("--verbose", action="store_true")
     args = parser.parse_args()
-
     if args.version:
         print_version()
         exit(0)
-    if not args.name:
-        exit("path is required")
 
     _setup_logging(args.verbose)
-
     try:
         package = _create_package(args)
         package.print_success()
         exit(0)
     except CppError as e:
         exit(e)
+    except KeyboardInterrupt:
+        print()
+        exit(1)
 
 
 def build_package():
@@ -244,51 +177,6 @@ def build_package():
         elif not (path / "setup.py").is_file():
             exit(f"{str(path / 'setup.py')} does not exist")
         _build_package(Path(args.path))
-    except CppError as e:
-        exit(e)
-
-
-def create_venv():
-    parser = argparse.ArgumentParser(
-        description="create a virtual environment for the given directory"
-    )
-    parser.add_argument(
-        "path",
-        default=".",
-        help="Path where venv should be created (usually the root of a package next to setup.py)",
-    )
-    parser.add_argument(
-        "--python",
-        default=sys.executable,
-        help="Python binary this venv should be associated with",
-    )
-    parser.add_argument("--verbose", action="store_true")
-
-    args = parser.parse_args()
-    _setup_logging(args.verbose)
-    try:
-        path = Path(args.path).resolve()
-        mkdir(path)
-        venv = _create_venv(path, path.name, args.python)
-        symlink_venv(path)
-        if (path / "requirements.txt").exists():
-            logging.info("Installing from requirements.txt file")
-            _run(
-                [
-                    venv / "bin" / "python",
-                    "-m",
-                    "pip",
-                    "install",
-                    "-r",
-                    path / "requirements.txt",
-                ]
-            )
-        printblue("A new virtual environment has been created!")
-        print()
-        print(
-            f"Run cd {str(path)!r} then type `source venv/bin/activate` "
-            "to activate it. When you are finished type `deactivate` to exit the environment."
-        )
     except CppError as e:
         exit(e)