+++ createpythonpackage/templates/entrypoint.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+__version__ = "{{version}}"
+
+
+def main():
+    print("TODO")
+
+
+if __name__ == "__main__":
+    main()