+++ createpythonpackage/templatestrs.py
@@ -1,0 +1,22 @@
+readme = """# {name}
+{description}
+"""
+
+gitignore = """*.pyc
+venv/
+activate-venv
+bin/
+build/
+develop-eggs/
+dist/
+eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+*.egg-info/
+.installed.cfg
+*.egg
+__pypackages__
+"""