+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # `create-python-package` Changelog
+0.2.0.0
+* Remove all entrypoints but `create-python-package`
+* Add interactive choices to command line
+* Support multiple licenses
+* Support venv, Pipenv, and poetry
 
 0.1.0.0
 * add makefile